Fix metadata for subjects page
